+++ util/package.go
@@ -60,7 +60,6 @@ type Package struct {
 	Categories      []string         `config:"categories" json:"categories"`
 	Release         string           `config:"release,omitempty" json:"release,omitempty"`
 	Conditions      *Conditions      `config:"conditions,omitempty" json:"conditions,omitempty" yaml:"conditions,omitempty"`
-	Requirement     Requirement      `config:"requirement" json:"requirement"`
 	Screenshots     []Image          `config:"screenshots,omitempty" json:"screenshots,omitempty" yaml:"screenshots,omitempty"`
 	Assets          []string         `config:"assets,omitempty" json:"assets,omitempty" yaml:"assets,omitempty"`
 	ConfigTemplates []ConfigTemplate `config:"config_templates,omitempty" json:"config_templates,omitempty" yaml:"config_templates,omitempty"`
@@ -93,18 +92,9 @@ type ConfigTemplate struct {
 	Multiple    *bool   `config:"multiple" json:"multiple,omitempty" yaml:"multiple,omitempty"`
 }
 
-type Requirement struct {
-	Kibana ProductRequirement `config:"kibana" json:"kibana,omitempty" yaml:"kibana"`
-}
-
 type Conditions struct {
 	KibanaVersion    string `config:"kibana.version,omitempty" json:"kibana.version,omitempty" yaml:"kibana.version,omitempty"`
 	kibanaConstraint *semver.Constraints
-}
-
-type ProductRequirement struct {
-	Versions    string `config:"versions,omitempty" json:"versions,omitempty" yaml:"versions,omitempty"`
-	semVerRange *semver.Constraints
 }
 
 type Version struct {
@@ -190,20 +180,10 @@ func NewPackage(basePath string) (*Package, error) {
 
 	p.Downloads = []Download{NewDownload(*p, "tar")}
 
-	// If the new conditions are used, select them over the requirements
 	if p.Conditions != nil && p.Conditions.KibanaVersion != "" {
 		p.Conditions.kibanaConstraint, err = semver.NewConstraint(p.Conditions.KibanaVersion)
 		if err != nil {
 			return nil, errors.Wrapf(err, "invalid Kibana versions range: %s", p.Conditions.KibanaVersion)
-		}
-		// TODO: remove legacy part
-	} else if p.Requirement.Kibana.Versions != "" {
-		p.Conditions = &Conditions{
-			KibanaVersion: p.Requirement.Kibana.Versions,
-		}
-		p.Conditions.kibanaConstraint, err = semver.NewConstraint(p.Requirement.Kibana.Versions)
-		if err != nil {
-			return nil, errors.Wrapf(err, "invalid Kibana versions range: %s", p.Requirement.Kibana.Versions)
 		}
 	}
 
@@ -268,7 +248,7 @@ func (p *Package) HasCategory(category string) bool {
 func (p *Package) HasKibanaVersion(version *semver.Version) bool {
 
 	// If the version is not specified, it is for all versions
-	if p.Conditions == nil || version == nil {
+	if p.Conditions == nil || version == nil || p.Conditions.kibanaConstraint == nil {
 		return true
 	}
 
@@ -354,13 +334,6 @@ func (p *Package) Validate() error {
 
 	if p.Description == "" {
 		return fmt.Errorf("no description set")
-	}
-
-	if p.Requirement.Kibana.Versions != "" {
-		_, err := semver.NewConstraint(p.Requirement.Kibana.Versions)
-		if err != nil {
-			return fmt.Errorf("invalid Kibana versions: %s, %s", p.Requirement.Kibana.Versions, err)
-		}
 	}
 
 	for _, c := range p.Categories {